+++ src/vector.rs
